requirements.txt
fastapi
uvicorn
openai>=1.40.0
requests
python-dotenv